+++ Paradigm/Changes.md
@@ -28,6 +28,7 @@
 - Removed certain unnecessary transformations by adding a check for identity transforms in 
   **LinearTransformable** and **AffineTransformable**.
 - Slightly refactored **RelativePoint** mapping logic
+- **Bounds**`.relativize(...)` now accepts any kind of **DoubleVector** instead of just a **Point**
 
 ## v1.5 - 22.01.2024
 This update contains some major refactoring in the following classes / concepts: